compile updates for clang

+++ src/maidsafe/common/tools/address_space_tool.cc
@@ -143,7 +143,8 @@ void Test::AddNode(bool good) {
 
     int group_common_leading_bits{GroupCommonLeadingBits(group_size)};
     int candidate_common_leading_bits{CandidateCommonLeadingBits(node_id, group_size)};
-    if (candidate_common_leading_bits < group_common_leading_bits + config_.leeway)
+    if (candidate_common_leading_bits <
+        static_cast<int>(group_common_leading_bits + config_.leeway))
       return DoAddNode(node_id, good, attempts);
   }
 }
